pyproject.toml
[project]
name = "toy-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "sqlmodel[async]>=0.0.22",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.1",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
]
